+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include README.md CHANGELOG.md LICENSE.txt utils
+include README.md CHANGELOG.md LICENSE.txt
